custom hook useFileImage
