+++ src/pages/ProductsConfig.jsx
@@ -1,52 +1,17 @@
-import { useRef, useState } from 'react'
+import { useRef } from 'react'
 import styled from 'styled-components'
 import sinfoto from '../assets/sinfoto_.png'
 import { BtnOpe } from '../components/BtnOpe'
 import { FaImage } from 'react-icons/fa6'
 import { Formulario } from '../components/Form'
+import { useFileImage } from '../hook/useFileImage'
 
 export const ProductsConfig = () => {
   const ref = useRef(null)
-  const [fileUrl, setFileUrl] = useState(sinfoto)
-  const [file, setFile] = useState([])
-  const [stateImage, setStateImage] = useState(false)
+  const { fileUrl, uploadImageStorage, stateImage, stateFormImage } = useFileImage({ sinfoto })
 
   const openImage = () => {
     ref.current.click()
-  }
-
-  const uploadImageStorage = (e) => {
-    // carga local
-    const fileLocal = e.target.files
-    const fileReaderLocal = new FileReader()
-    fileReaderLocal.readAsDataURL(fileLocal[0])
-    const typeImage = e.target.files[0]
-
-    if (!typeImage.type.includes('image/png')) return null
-    if (fileReaderLocal && fileLocal && fileLocal.length) {
-      fileReaderLocal.onload = function load () {
-        setFileUrl(fileReaderLocal.result)
-      }
-
-      // preparar img para el storage
-
-      const fileList = e.target.files
-      const fileReader = new FileReader()
-      fileReader.readAsArrayBuffer(fileList[0])
-      fileReader.onload = function () {
-        const imagenData = fileReader.result
-        setFile(imagenData)
-      }
-    }
-  }
-
-  const stateFormImage = () => {
-    const img = file.length
-    if (img !== 0) {
-      setStateImage(false)
-    } else {
-      setStateImage(true)
-    }
   }
 
   return (
